Doe inheritance met bbcode.
Op deze manier kun je bbcode groeperen op wat het doet in een bestand.
Het verminderd ook de complexiteit van specifieke email of light mode
tags.
