+++ lib/controller/api/ApiForumController.class.php
@@ -9,6 +9,7 @@ use CsrDelft\model\forum\ForumPostsModel;
 use CsrDelft\model\ProfielModel;
 use CsrDelft\model\security\LoginModel;
 use CsrDelft\view\bbcode\CsrBB;
+use CsrDelft\view\bbcode\LightCsrBB;
 use Exception;
 use Jacwright\RestServer\RestException;
 
@@ -71,7 +72,7 @@ class ApiForumController {
 
 		foreach ($posts as $post) {
 			$post->uid_naam = ProfielModel::getNaam($post->uid, 'civitas');
-			$post->tekst = CsrBB::parseLight($post->tekst);
+			$post->tekst = LightCsrBB::parseLight($post->tekst);
 		}
 
 		return array('data' => $posts);
